split CSV Grid into a standalone frame extension
Demonstrate the Application/Extension split using the CSV example:

  - cocalc/csv-editor   (Application)  -- file-type scaffold, sync
                                           config, default layout,
                                           Raw Data frame
  - cocalc/csv-grid-frame (Extension)  -- adds the Grid frame type
                                           cocalc/csv-grid, targeting
                                           cocalc/csv-editor

Wire frame extensions into EditorSpec: toEditorSpec() now merges
frames from extensionRegistry.getFrameCandidatesForEditor(editorId)
alongside native frames and the editor's own frames. Editor's own
frames win conflicts.

Both builtins ship together, so defaultLayout can still reference
cocalc/csv-grid; the frame extension is always loaded by the time
the editor spec is materialized (lazy, on first file open).
