+++ src/packages/sdk/builtin/csv-grid-frame/manifest.json
@@ -1,0 +1,5 @@
+{
+  "id": "cocalc/csv-grid-frame",
+  "version": "0.0.1",
+  "main": "extension.js"
+}
